+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add `paymentReaderCardDetails` parameter to `Transaction.sale`
+* Add error code `TRANSACTION_TAX_AMOUNT_IS_REQUIRED_FOR_AIB_SWEDISH` for attribute `taxAmount` in `transaction` key for AIB:Domestic transactions in Sweden.
 
 ## 6.2.0
 * Add `chargebackProtectionLevel ` to `Dispute` and `DisputeSearch`